TabFeaturesInternal -> BraveTabFeatures
